+++ R/two-column-layouts.R
@@ -174,7 +174,7 @@ gt_two_column_layout <- function(tables = NULL, output = "viewer",
 
     htmltools::save_html(html = double_tables, file = tempfile_)
 
-    # Saving an image requires the webshot package; if it's
+    # Saving an image requires the webshot2 package; if it's
     # not present, stop with a message
     if (!rlang::is_installed("webshot2")) {
       stop("The `webshot2` package is required for saving images of gt tables.)",