+++ NEWS.md
@@ -1,3 +1,8 @@
+# gtExtras 0.3.9
+- Renamed colors arg in `merge_and_stack()` to be 'palette'
+- Renamed colors arg in `gt_plt_bullet()` to be 'palette'
+- Renamed pal arg in `gt_plt_sparkline()` to be 'palette'
+
 # gtExtras 0.3.8
 - Renamed `colors` argument in `gt_win_loss() to `palette`
 - Added NA handling to all `fontawesome::fa()` functions, ie `gt_fa_rank_change()`, `gt_fa_repeats()`, `gt_fa_column()`